+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 - 再読み込みの後じゃないと good! が出ていかないことがある。
 - dev container?
 - report-pt! - 同人に複数票が入れられる。
+- colorful buttons.
+- 1:1 -> 2:1
+
+## v2.6-SNAPSHOT
+
+## v2.6.580 / 2024-05-30
+### Added
+- update button to `points received`.
+### Changed
+- radio to buttons
 
 ## v2.6.578 / 2024-05-30
 ### Added
@@ -13,9 +23,7 @@
   :middleware [(if (:dev env) identity middleware/wrap-restricted)
   ...
 ```
-- points sent/received
-### Todo
-- radio よりボタンが良くないか？
+- points sent
 
 
 ## v2.5.570 / 2024-05-30